Add logic to handle when multiple datasets are loaded via 51x script.

+++ data-processor/scripts/90-generate-site.sh
@@ -12,8 +12,10 @@ rm -rf site
 cp -r ../website/dist/map4sci/ site
 cp -r $OUT/site-data/* site/assets
 
-cat > site/assets/datasets/index.json << EOF
+if [ ! -e $OUT/site-data/datasets/index.json ]; then
+  cat > site/assets/datasets/index.json << EOF
 [
   {"id": "$DATASET", "name": "$DATASET", "dir": "assets/datasets/$DATASET"}
 ]
 EOF
+fi
